Add attachment on exam

+++ database/migrations/2018_10_18_170034_create_pmb_hasil_table.php
@@ -21,7 +21,7 @@ class CreatePmbHasilTable extends Migration
             $table->string('kode_jurusan');
             $table->string('kode_soal');
             $table->string('kode_kelas');
-            $table->float('nilai_angka');
+            $table->integer('nilai_angka');
             $table->timestamps();
         });
     }
